refactor(core): refactor types and main class name
- Moves types into root src
- Rename main class to TerraformCloud
