+++ test/mocks/index.ts
@@ -1,0 +1,4 @@
+import AccountMock from './AccountMock'
+import RunMock from './RunMock'
+
+export { AccountMock, RunMock }
